Squash abstract database into simply base database
