+++ data_diff/abcs/database_types.py
@@ -268,76 +268,6 @@ class AbstractDialect(ABC):
         """Ensure that the expression is comparable in ``IS DISTINCT FROM``."""
 
 
-from typing import TypeVar, Generic
-
-T_Dialect = TypeVar("T_Dialect", bound=AbstractDialect)
-
-
-class AbstractDatabase(Generic[T_Dialect]):
-    @property
-    @abstractmethod
-    def dialect(self) -> T_Dialect:
-        "The dialect of the database. Used internally by Database, and also available publicly."
-
-    @classmethod
-    @abstractmethod
-    def load_mixins(cls, *abstract_mixins) -> type:
-        "Extend the dialect with a list of mixins that implement the given abstract mixins."
-
-    @property
-    @abstractmethod
-    def CONNECT_URI_HELP(self) -> str:
-        "Example URI to show the user in help and error messages"
-
-    @property
-    @abstractmethod
-    def CONNECT_URI_PARAMS(self) -> List[str]:
-        "List of parameters given in the path of the URI"
-
-    @abstractmethod
-    def _query(self, sql_code: str) -> list:
-        "Send query to database and return result"
-
-    @abstractmethod
-    def query_table_schema(self, path: DbPath) -> Dict[str, tuple]:
-        """Query the table for its schema for table in 'path', and return {column: tuple}
-        where the tuple is (table_name, col_name, type_repr, datetime_precision?, numeric_precision?, numeric_scale?)
-
-        Note: This method exists instead of select_table_schema(), just because not all databases support
-              accessing the schema using a SQL query.
-        """
-
-    @abstractmethod
-    def select_table_unique_columns(self, path: DbPath) -> str:
-        "Provide SQL for selecting the names of unique columns in the table"
-
-    @abstractmethod
-    def query_table_unique_columns(self, path: DbPath) -> List[str]:
-        """Query the table for its unique columns for table in 'path', and return {column}"""
-
-    @abstractmethod
-    def _process_table_schema(
-        self, path: DbPath, raw_schema: Dict[str, tuple], filter_columns: Sequence[str], where: str = None
-    ):
-        """Process the result of query_table_schema().
-
-        Done in a separate step, to minimize the amount of processed columns.
-        Needed because processing each column may:
-        * throw errors and warnings
-        * query the database to sample values
-
-        """
-
-    @abstractmethod
-    def close(self):
-        "Close connection(s) to the database instance. Querying will stop functioning."
-
-    @property
-    @abstractmethod
-    def is_autocommit(self) -> bool:
-        "Return whether the database autocommits changes. When false, COMMIT statements are skipped."
-
-
 class AbstractTable(ABC):
     @abstractmethod
     def select(self, *exprs, distinct=False, **named_exprs) -> "AbstractTable":
